Updated db models with default data
Added a `code` column (which is a unique natural key) to all database
models that are initialized with default data. This removes the need to
manually define the primary keys for the default data for these models.
Manually setting primary keys when using an identity column generated by
default by the database could be problematic when using MS SQL Server.
You may have to use `SET IDENTITY_INSERT table ON`. The natural unique
code is also self-documenting in its use compared to an integer.