Updated creation of default database data and initialization
Changed to define the default data as dictionaries rather than global
SQLAlchemy ORM models. The global models are problematic when having to
build test fixtures that expose an initialized database with default data.
If the global models have been added to a SQLAlchemy session and committed
to a database subsequent use of these models in other sessions will not
persist them to a new database. This leads to tests passing when running
them in isolation but failing when running the entire test suite. The new
setup will avoid this issue.