Added fixture reset_log_handlers
Ensure that log handlers are reset after every test.

This fixture makes sure that tests that use the `capsys` fixture do not
cause failures for other tests downstream. When `capsys` is torn down at
the end of a test it closes the stream it captures and thus subsequent
tests that tries to write to stdout or stderr will try to write to a closed
stream, which will raise: "ValueError: I/O operation on closed file." and
fail the test.

The CLI tests write to stdout using the logging module and `reset_log_handlers`
ensures that these tests can run correctly even after tests that use `capsys`
have run. All tests now pass when run together in a single test session.